test(gui): rewrite the Settings panel suite from the panel's own controls
Family B. 35 cases over 1,287 effective lines become 40 over 1,439, and the extra
lines are coverage rather than prose: six catalog propositions had no case at all.
The list's row count in its own section header, the Source cell's Mine/Factory text,
the read-only axis cells being disabled CONTROLS rather than text, the ' (mine)'
suffix on a retuned preset, the action row staying pinned while the list shrinks, and
the display precision of a clamped std -- the one the user reads as '15' directly
under a line saying the value must exceed 15.

Four of those were unfalsifiable the way the old suite addressed things. Every id
this panel exposes uses the '###' form, so an item is found whichever text it drew;
the drawn text is only reachable through ItemInfo().DebugLabel, which is how the
count, the Mine/Factory word and the (mine) suffix are now asserted. The clamped std
is read by activating its input, the only moment its formatted text exists anywhere.
The read-only axis cells share their labels across three rows, so no wildcard search
has a unique match at all -- their ids are derived from the ID stack the panel
builds, seeded on a window taken at runtime and checked against a known item's id
before being trusted.

What the layer cannot answer is recorded rather than faked: the status line is a
TextWrapped submitted with id 0, so 'Adjusted to ...', 'No changes to save.' and
'Your personal defaults were updated.' are invisible here. Each of those
propositions keeps the half that has an observable -- the stored value, the warning
cell, the file -- and none of them pretends to check the sentence.

The fixture folds every preset node back on the way out. ImGui keeps a TreeNode's
open flag in window storage that outlives every reset this suite performs, so a case
that unfolds one hands the next a library already open -- and the visual reference
scenes downstream capture exactly one preset unfolded. Measured at 16.8 dB on
presets_expanded and presets_warning before the teardown existed.

Also fixes a racy assertion in the run-lifecycle suite, reproduced here rather than
dismissed: Stop's intent is read as kStopping immediately after the click, but the
drain is offloaded and when it lands inside the same click the poller has already
advanced to kStopped. It fails as '[4] == [3]' under a real-time frame clock and
intermittently under --fixed-dt. The claim is now what the click actually
guarantees -- the intent leaves kRunning synchronously -- which still fails on a Stop
button that queued its intent behind the drain.
